// src/app/model/alert/weather-field.ts
/* eslint-disable @typescript-eslint/naming-convention */
export enum WeatherField {
  TEMP = 'TEMP',
  WIND = 'WIND',
  PRESSURE = 'PRESSURE',
  PRECIP = 'PRECIP',
  HUMIDITY = 'HUMIDITY',
  CLOUD = 'CLOUD',
  FEELS_LIKE = 'FEELS_LIKE',
  CHANCE_OF_RAIN = 'CHANCE_OF_RAIN',
  CHANCE_OF_SNOW = 'CHANCE_OF_SNOW',
  UV = 'UV',
}